moved Evaluation to it's own file
