+++ src/simulated_annealing/neighbor_move/mod.rs
@@ -1,5 +1,6 @@
 // I DO NOT KNOW WHAT THIS THING SHOULD BE NAMED YET
 pub mod neighbor_move_trait;
+pub mod evaluation;
 pub mod shift_in_route;
 pub mod evaluation_helper;
 pub mod shift_between_days;
